+++ e2e-tests/commands/insurance/complete-and-submit-how-was-the-contract-awarded-form.js
@@ -1,21 +1,22 @@
-import { radios } from '../../pages/shared';
-import { EXPORT_CONTRACT as EXPORT_CONTRACT_FIELD_IDS } from '../../constants/field-ids/insurance/export-contract';
-import { EXPORT_CONTRACT_FIELDS as FIELDS } from '../../content-strings/fields/insurance';
-
-const {
-  HOW_WAS_THE_CONTRACT_AWARDED: { AWARD_METHOD },
-} = EXPORT_CONTRACT_FIELD_IDS;
-
-const AWARD_METHOD_OPTIONS = FIELDS.HOW_WAS_THE_CONTRACT_AWARDED[AWARD_METHOD].OPTIONS;
-
 /**
  * completeAndSubmitHowWasTheContractAwardedForm
  * Complete and submit the "How was the contract awarded" form
+ * @param {Boolean} openTender: Award method as OPEN_TENDER
+ * @param {Boolean} negotiatedContract: Award method as NEGOTIATED_CONTRACT
+ * @param {Boolean} directAward: Award method as DIRECT_AWARD
+ * @param {Boolean} competitiveBidding: Award method as COMPETITIVE_BIDDING
+ * @param {Boolean} otherMethod: Award method as OTHER
+ * @param {String} otherMethodText: OTHER award method text
  */
-const completeAndSubmitHowWasTheContractAwardedForm = () => {
-  const selector = radios(AWARD_METHOD_OPTIONS.OPEN_TENDER.ID).option;
-
-  selector.label().click();
+const completeAndSubmitHowWasTheContractAwardedForm = ({ openTender, negotiatedContract, directAward, competitiveBidding, otherMethod, otherMethodText }) => {
+  cy.completeHowWasTheContractAwardedForm({
+    openTender,
+    negotiatedContract,
+    directAward,
+    competitiveBidding,
+    otherMethod,
+    otherMethodText,
+  });
 
   cy.clickSubmitButton();
 };