+++ e2e-tests/insurance/cypress/e2e/journeys/export-contract/agent-details/agent-details.spec.js
@@ -33,7 +33,7 @@ context(
 
         // go to the page we want to test.
         cy.startInsuranceExportContractSection({});
-        cy.completeAndSubmitHowWasTheContractAwardedForm();
+        cy.completeAndSubmitHowWasTheContractAwardedForm({});
         cy.completeAndSubmitAboutGoodsOrServicesForm({});
         cy.completeAndSubmitHowYouWillGetPaidForm({});
         cy.completeAndSubmitAgentForm({ isUsingAgent: true });