+++ e2e-tests/insurance/cypress/e2e/journeys/export-contract/agent/save-and-back.spec.js
@@ -22,7 +22,7 @@ context('Insurance - Export contract - Agent - Save and go back', () => {
 
       // go to the page we want to test.
       cy.startInsuranceExportContractSection({});
-      cy.completeAndSubmitHowWasTheContractAwardedForm();
+      cy.completeAndSubmitHowWasTheContractAwardedForm({});
       cy.completeAndSubmitAboutGoodsOrServicesForm({});
       cy.completeAndSubmitHowYouWillGetPaidForm({});
 
@@ -75,11 +75,8 @@ context('Insurance - Export contract - Agent - Save and go back', () => {
 
         cy.startInsuranceExportContractSection({});
 
-        // TODO: EMS-3665 - increment the value in command below instead.
-        cy.completeAndSubmitHowWasTheContractAwardedForm();
-
-        // go through 2 export contract forms.
-        cy.clickSubmitButtonMultipleTimes({ count: 2 });
+        // go through 3 export contract forms.
+        cy.clickSubmitButtonMultipleTimes({ count: 3 });
 
         cy.assertNoRadioOptionIsChecked();
       });
@@ -108,11 +105,8 @@ context('Insurance - Export contract - Agent - Save and go back', () => {
 
         cy.startInsuranceExportContractSection({});
 
-        // TODO: EMS-3665 - increment the value in command below instead.
-        cy.completeAndSubmitHowWasTheContractAwardedForm();
-
-        // go through 2 export contract forms.
-        cy.clickSubmitButtonMultipleTimes({ count: 2 });
+        // go through 3 export contract forms.
+        cy.clickSubmitButtonMultipleTimes({ count: 3 });
 
         cy.assertYesRadioOptionIsChecked();
       });