+++ e2e-tests/insurance/cypress/e2e/journeys/export-contract/declined-by-private-market/save-and-back.spec.js
@@ -23,7 +23,7 @@ context('Insurance - Export contract - Declined by private market - Save and go 
 
       // go to the page we want to test.
       cy.startInsuranceExportContractSection({});
-      cy.completeAndSubmitHowWasTheContractAwardedForm();
+      cy.completeAndSubmitHowWasTheContractAwardedForm({});
       cy.completeAndSubmitAboutGoodsOrServicesForm({});
       cy.completeAndSubmitHowYouWillGetPaidForm({});
       cy.completeAndSubmitPrivateMarketForm({ attemptedPrivateMarketCover: true });
@@ -76,11 +76,8 @@ context('Insurance - Export contract - Declined by private market - Save and go 
 
         cy.startInsuranceExportContractSection({});
 
-        // TODO: EMS-3665 - increment the value in command below instead.
-        cy.completeAndSubmitHowWasTheContractAwardedForm();
-
-        // go through 3 export contract forms.
-        cy.clickSubmitButtonMultipleTimes({ count: 3 });
+        // go through 4 export contract forms.
+        cy.clickSubmitButtonMultipleTimes({ count: 4 });
 
         cy.checkTextareaValue({
           fieldId: FIELD_ID,