+++ src/ui/server/controllers/insurance/export-contract/how-was-the-contract-awarded/index.ts
@@ -9,11 +9,12 @@ import getUserNameFromSession from '../../../../helpers/get-user-name-from-sessi
 import constructPayload from '../../../../helpers/construct-payload';
 import generateValidationErrors from './validation';
 import { sanitiseData } from '../../../../helpers/sanitise-data';
-import { Request, Response } from '../../../../../types';
+import mapAndSave from '../map-and-save/export-contract';
+import { ObjectType, Request, Response } from '../../../../../types';
 
 const {
   INSURANCE_ROOT,
-  EXPORT_CONTRACT: { ABOUT_GOODS_OR_SERVICES },
+  EXPORT_CONTRACT: { ABOUT_GOODS_OR_SERVICES, HOW_WAS_THE_CONTRACT_AWARDED_SAVE_AND_BACK: SAVE_AND_BACK },
   PROBLEM_WITH_SERVICE,
 } = INSURANCE_ROUTES;
 
@@ -38,9 +39,10 @@ export const FIELD_IDS = [AWARD_METHOD, OTHER_AWARD_METHOD];
 /**
  * pageVariables
  * Page fields and "save and go back" URL
+ * @param {Number} Application reference number
  * @returns {Object} Page variables
  */
-export const pageVariables = () => ({
+export const pageVariables = (referenceNumber: number) => ({
   FIELDS: {
     AWARD_METHOD: {
       ID: AWARD_METHOD,
@@ -51,7 +53,7 @@ export const pageVariables = () => ({
       ...FIELDS.HOW_WAS_THE_CONTRACT_AWARDED[OTHER_AWARD_METHOD],
     },
   },
-  SAVE_AND_BACK_URL: '#',
+  SAVE_AND_BACK_URL: `${INSURANCE_ROOT}/${referenceNumber}${SAVE_AND_BACK}`,
 });
 
 /**
@@ -68,14 +70,17 @@ export const get = (req: Request, res: Response) => {
     return res.redirect(PROBLEM_WITH_SERVICE);
   }
 
+  const { referenceNumber } = application;
+
   return res.render(TEMPLATE, {
     ...insuranceCorePageVariables({
       PAGE_CONTENT_STRINGS,
       BACK_LINK: req.headers.referer,
     }),
-    ...pageVariables(),
+    ...pageVariables(referenceNumber),
     CONDITIONAL_OTHER_METHOD_HTML,
     userName: getUserNameFromSession(req.session.user),
+    submittedValues: application.exportContract,
   });
 };
 
@@ -100,18 +105,45 @@ export const post = async (req: Request, res: Response) => {
   const validationErrors = generateValidationErrors(payload);
 
   if (validationErrors) {
+    const sanitised = sanitiseData(payload) as ObjectType;
+
+    /**
+     * Map the payload into an AWARD_METHOD object structure with an id property.
+     * Otherwise, the nunjucks template needs 2x conditions.
+     */
+    const submittedValues = {
+      ...sanitised,
+      [AWARD_METHOD]: {
+        id: sanitised[AWARD_METHOD],
+      },
+    };
+
     return res.render(TEMPLATE, {
       ...insuranceCorePageVariables({
         PAGE_CONTENT_STRINGS,
         BACK_LINK: req.headers.referer,
       }),
-      ...pageVariables(),
+      ...pageVariables(referenceNumber),
       CONDITIONAL_OTHER_METHOD_HTML,
       userName: getUserNameFromSession(req.session.user),
-      submittedValues: sanitiseData(payload),
+      submittedValues,
       validationErrors,
     });
   }
 
-  return res.redirect(`${INSURANCE_ROOT}/${referenceNumber}${ABOUT_GOODS_OR_SERVICES}`);
+  try {
+    // save the application
+
+    const saveResponse = await mapAndSave.exportContract(payload, application, validationErrors);
+
+    if (!saveResponse) {
+      return res.redirect(PROBLEM_WITH_SERVICE);
+    }
+
+    return res.redirect(`${INSURANCE_ROOT}/${referenceNumber}${ABOUT_GOODS_OR_SERVICES}`);
+  } catch (err) {
+    console.error('Error updating application - export contract - how was the contract awarded %O', err);
+
+    return res.redirect(PROBLEM_WITH_SERVICE);
+  }
 };