+++ src/ui/server/controllers/insurance/export-contract/map-submitted-data/export-contract/index.ts
@@ -1,23 +1,45 @@
 import FIELD_IDS from '../../../../../constants/field-ids/insurance/export-contract';
 import getCountryByIsoCode from '../../../../../helpers/get-country-by-iso-code';
 import { objectHasProperty } from '../../../../../helpers/object';
+import { isEmptyString } from '../../../../../helpers/string';
 import { Country, RequestBody } from '../../../../../../types';
 
 const {
+  HOW_WAS_THE_CONTRACT_AWARDED: { AWARD_METHOD },
   ABOUT_GOODS_OR_SERVICES: { FINAL_DESTINATION, FINAL_DESTINATION_KNOWN },
 } = FIELD_IDS;
 
 /**
  * mapSubmittedData
- * Map About goods or services fields
- * if FINAL_DESTINATION is provided, map as the country ISO code.
- * if FINAL_DESTINATION_KNOWN is false, delete FINAL_DESTINATIN
+ * Map "export contract" related data for the following forms:
+ * - "How was the contract awarded"
+ * - "About goods or services"
+ * 1) if AWARD_METHOD is an empty string, delete the field.
+ * 2) if AWARD_METHOD is provided, map as a KeystoneJS connect object with ID relationship.
+ * 3) if FINAL_DESTINATION is provided, map as the country ISO code.
+ * 4) if FINAL_DESTINATION_KNOWN is false, delete FINAL_DESTINATION.
  * @param {RequestBody} formBody
  * @param {Array<Country>} countries
  * @returns {Object} populatedData
  */
 const mapSubmittedData = (formBody: RequestBody, countries?: Array<Country>): object => {
   const populatedData = formBody;
+
+  /**
+   * If AWARD_METHOD is an empty string,
+   * delete the field.
+   */
+  if (isEmptyString(formBody[AWARD_METHOD])) {
+    delete populatedData[AWARD_METHOD];
+  }
+
+  if (objectHasProperty(populatedData, AWARD_METHOD)) {
+    populatedData[AWARD_METHOD] = {
+      connect: {
+        id: formBody[AWARD_METHOD],
+      },
+    };
+  }
 
   const hasFinalDestination = objectHasProperty(formBody, FINAL_DESTINATION);